* Use `uint8_t __ARDUINO_CI_SFR_MOCK[1024]` instead of `long long`. * Set bit order based on `SPISettings` in `beginTransaction()` and use `SPCR` and `DORD` to test it.

+++ cpp/arduino/Godmode.cpp
@@ -114,4 +114,4 @@ SPIClass SPI = SPIClass(&GODMODE()->spi.dataIn, &GODMODE()->spi.dataOut);
 // defined in Wire.h
 TwoWire Wire = TwoWire();
 
-volatile long long __ARDUINO_CI_SFR_MOCK[1024];
+volatile uint8_t __ARDUINO_CI_SFR_MOCK[1024];
